デバッグ用コマンドを追加
